+++ worker/pool.go
@@ -5,39 +5,6 @@ import (
 	"github.com/berachain/offchain-sdk/log"
 )
 
-// PoolConfig is the configuration for a pool.
-type PoolConfig struct {
-	// Name is the name of the pool.
-	Name string
-	// PrometheusPrefix is the prefix for the prometheus metrics.
-	PrometheusPrefix string
-	// MinWorkers is the minimum number of workers that the resizer will
-	// shrink the pool down to .
-	MinWorkers int
-	// MaxWorkers is the maximum number of workers that can be active
-	// at the same time.
-	MaxWorkers int
-	// ResizingStrategy is the methodology used to resize the number of workers
-	// in the pool.
-	ResizingStrategy string
-	// MaxQueuedJobs is the maximum number of jobs that can be queued
-	// before the pool starts rejecting jobs.
-	MaxQueuedJobs int
-}
-
-// DefaultPoolConfig is the default configuration for a pool.
-func DefaultPoolConfig() *PoolConfig {
-	return &PoolConfig{
-		Name:             "default",
-		PrometheusPrefix: "default",
-		MinWorkers:       4,  //nolint:gomnd // it's ok.
-		MaxWorkers:       32, //nolint:gomnd // it's ok.
-		ResizingStrategy: "balanced",
-		MaxQueuedJobs:    100, //nolint:gomnd // it's ok.
-	}
-}
-
-// Pool is a wrapper around a pond.WorkerPool. We use this to add logging
 // and other functionality to the pool.
 type Pool struct {
 	name   string
